add time resource
When configured with a duration (in the source), the time resource will report
new a new version if it's been that long since the previous version.

This allows builds configured on arbitrary intervals (say, months) to be
guaranteed to build, rather than relying on intervals scheduled by some
ephemeral component.